Add files via upload
The goal of a heart failure prediction project is to develop a machine learning model that can accurately predict the likelihood of an individual experiencing heart failure based on a set of input features. Heart failure is a serious medical condition where the heart is unable to pump blood effectively to meet the body's needs, leading to symptoms such as shortness of breath, fatigue, and fluid retention.

By leveraging data such as demographic information, medical history, lifestyle factors, and clinical measurements, the aim is to create a predictive model that can assist healthcare professionals in identifying individuals who are at a higher risk of developing heart failure. Early detection and intervention can lead to better management of the condition, improved outcomes for patients, and potentially reduced healthcare costs.

The heart failure prediction project typically involves several key steps, including data collection and preprocessing, feature selection and engineering, model training and evaluation, and deployment of the predictive model in a clinical setting. The ultimate objective is to build a robust and accurate predictive tool that can aid in the early identification and management of individuals at risk of heart failure, ultimately contributing to better patient care and outcomes.